+++ src/menu.rs
@@ -340,6 +340,7 @@ const VIEW_EDITOR: &[Item] = &[
     Item::leaf("menu.item.view.inlay_hints", "view.inlay_hints", ""),
     Item::leaf("menu.item.view.sticky_scroll", "view.sticky_scroll", ""),
     Item::leaf("menu.item.view.minimap", "view.minimap", ""),
+    Item::leaf("menu.item.view.highlight_word", "view.highlight_word", ""),
     Item::leaf("menu.item.view.spellcheck", "view.spellcheck", ""),
     SEP,
     Item::leaf("menu.item.view.auto_pair", "view.auto_pair", ""),